feat: createTrackedSelector
