+++ package.json
@@ -1,7 +1,10 @@
 {
   "name": "react-tracked",
   "description": "Simple and fast global state with React Context. Eliminate unnecessary re-renders without hassle.",
-  "version": "1.5.0",
+  "version": "1.6.0-beta.1",
+  "publishConfig": {
+    "tag": "beta"
+  },
   "author": "Daishi Kato",
   "repository": {
     "type": "git",
