+++ src/index.ts
@@ -1,3 +1,4 @@
 export { createContainer } from './createContainer';
+export { createTrackedSelector } from './createTrackedSelector';
 export { memo } from './memo';
 export { getUntrackedObject } from 'proxy-compare';
